UI data sorting and sim end detection.
The UI now detects when a simulation has functionally ended even if it hasn't reached the number of time steps in the sim.

Simulations in the left hand panel and agent information within the active content panel are now sorted by simID and sequential ID respectively. This makes it easier to find relevant information.